Rename project topbar components
