+++ app/src/components/ProjectTopbarMenuItem.svelte
